Add option to connect/disconnect based on user status
This is set to false by default now, so users' clients will connect based on messages and disconnect based on the timeout from the last message sent

Also added better message queueing in the even that a client is starting up. It will now wait until the user's irc client says it is connected before attempting to send messages to the channel.